% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_actioncard}
\alias{send_actioncard}
\alias{send_actioncard_multi}
\title{Send actioncard}
\usage{
send_actioncard(title, text, singleTitle, singleURL,
  btnOrientation = FALSE, hideAvatar = FALSE, prepare = FALSE)

send_actioncard_multi(title, text, btns, btnOrientation = FALSE,
  hideAvatar = FALSE, prepare = FALSE)
}
\arguments{
\item{title}{title text}

\item{text}{markdown text}

\item{singleTitle}{title of actioncard}

\item{singleURL}{URL of actioncard}

\item{prepare}{return json request body, for debug use only}

\item{btns}{a nested list with each element containing two fields: "title" and "actionURL"}

\item{...}{url or access_token}
}
\description{
Send actioncard
}
