% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_feedcard}
\alias{send_feedcard}
\title{Send feedcard}
\usage{
send_feedcard(links, prepare = FALSE)
}
\arguments{
\item{links}{a nested list with each element containing three fields: "title", "messageURL" and "picURL"}

\item{prepare}{return json request body, for debug use only}

\item{...}{url or access_token}
}
\description{
Send feedcard
}
