% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_link}
\alias{send_link}
\title{Send link}
\usage{
send_link(title, text, messageUrl, picUrl = NULL, prepare = FALSE)
}
\arguments{
\item{title}{title text}

\item{text}{plain text}

\item{messageUrl}{URL to redirect to}

\item{picUrl}{URL of the picture to display}

\item{prepare}{return json request body, for debug use only}

\item{...}{url or access_token}
}
\description{
Send link
}
