% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_markdown}
\alias{send_markdown}
\title{Send markdown}
\usage{
send_markdown(title, text, atMobiles = NULL, isAtAll = FALSE,
  prepare = FALSE)
}
\arguments{
\item{title}{title text}

\item{text}{markdown text}

\item{atMobiles}{vector of mobile phone numbers to @}

\item{isAtAll}{if TRUE, @ all users}

\item{prepare}{return json request body, for debug use only}

\item{...}{url or access_token}
}
\description{
Send markdown
}
