% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send_text}
\alias{send_text}
\title{Send plain text}
\usage{
send_text(content, atMobiles = NULL, isAtAll = FALSE,
  prepare = FALSE)
}
\arguments{
\item{content}{plain text}

\item{atMobiles}{vector of mobile phone numbers to @}

\item{isAtAll}{if TRUE, @ all users}

\item{prepare}{return json request body, for debug use only}

\item{...}{url or access_token}
}
\description{
Send plain text
}
