% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot.R
\name{set_dingrobot}
\alias{set_dingrobot}
\title{Set Ding Robot webhook address}
\usage{
set_dingrobot(url = NULL, access_token = NULL, secret = NULL)
}
\arguments{
\item{url}{webhook URL address}

\item{access_token}{token to generate webhook}

\item{secret}{secret key if signing required}
}
\description{
Set Ding Robot webhook address
}
